refactor(qqbot): 独立指令和普通消息处理逻辑
